lacrei_saude/pyproject.toml
[project]
name = "lacrei-saude"
version = "0.1.0"
description = "Desafio de codigo para a Lacrei Saude"
authors = [
    {name = "jess",email = "70455379+jesslourenco@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "bleach (>=6.2.0,<7.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
